requirements-cloudflare.txt
# Requirements spécifiques pour Cloudflare Workers
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0

# Packages spécifiques pour Cloudflare Workers
cloudflare==2.11.1
wrangler==3.0.0
